Refactor confirm and delete confirmation modals into a single reusable component. Update styles. Remove unnecessary handlers
