+++ src/components/AddNewProduct/AddNewProduct.js
@@ -1,14 +1,14 @@
 import './AddNewProduct.css'
 import Button from '../Button/Button'
 
-const AddNewProduct = ({ handleOpenModal }) => {
+const AddNewProduct = ({ setOpenModal }) => {
   return (
     <section className='add-new-product-container'>
       <h2 className='inventory-heading'>Inventory</h2>
       <Button
         name='New product'
         className='new-product-button'
-        onClick={handleOpenModal}
+        onClick={() => setOpenModal(true)}
       />
     </section>
   )
