--- src/components/DeleteConfirmationModal/DeleteConfirmationModal.css
@@ -1,74 +1,0 @@
-.delete-confirmation-modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  backdrop-filter: blur(2px);
-  background-color: rgba(163, 163, 163, 0.2);
-  z-index: 1000;
-}
-
-.delete-confirmation-modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: #fff;
-  padding: 1.25rem 2rem;
-  z-index: 1000;
-  border-radius: 0.75rem;
-  line-height: 1.5rem;
-}
-
-.delete-confirmation-heading::before {
-  content: url('../../assets/icons/blue-warning.svg');
-  display: inline-block;
-  vertical-align: middle;
-  padding-right: 0.5rem;
-}
-
-.delete-confirmation-heading {
-  font-weight: 700;
-  font-size: 1.25rem;
-  margin-top: 1rem;
-}
-
-.delete-buttons-container {
-  display: flex;
-  gap: 2rem;
-  justify-content: center;
-  margin-top: 1.75rem;
-}
-
-.cancel-button,
-.delete-button {
-  font-weight: 500;
-  font-size: 1.1rem;
-  border-radius: 0.75rem;
-  border: 2px solid #1162dd;
-  padding-top: 10px;
-  padding-bottom: 12px;
-  cursor: pointer;
-}
-
-.cancel-button {
-  color: #1162dd;
-  background-color: #fff;
-  padding-left: 2rem;
-  padding-right: 2rem;
-}
-
-.cancel-button:hover {
-  background-color: #deebff;
-}
-.delete-button {
-  color: #fff;
-  background-color: #1162dd;
-  padding-left: 2rem;
-  padding-right: 2rem;
-}
-
-.delete-button:hover {
-  background-color: #5c9dff;
-}
