+++ src/components/MainSectionContent/MainSectionContent.css
@@ -5,17 +5,17 @@
 }
 
 main {
-  margin-left: 2.188rem;
-  margin-right: 2rem;
-  width: 96%;
+  padding-bottom: 2rem;
 }
 
 .products-container {
   box-shadow: 0px 0px 4px 1px #dadada;
   border-radius: 12px;
-  height: 40rem;
+  height: 36rem;
   max-height: 49rem;
   overflow: scroll;
   scroll-behavior: smooth;
   overflow-x: hidden;
+  margin-left: 2rem;
+  margin-right: 2rem;
 }
