+++ src/components/ProductFormModal/ProductFormModal.js
@@ -14,7 +14,7 @@ import Select from 'react-select'
 import './ProductFormModal.css'
 
 const ProductFormModal = ({
-  handleCloseModal,
+  setOpenModal,
   handleAddNewProduct,
   handleUpdateExistingProduct,
   handlePartialEdit,
@@ -30,18 +30,17 @@ const ProductFormModal = ({
     control,
     formState: { errors },
   } = useForm({ defaultValues: dataToEdit ? dataToEdit : defaultFormValues })
-
   const imageProps = watch(['imgUrl', 'name'])
 
   const handleData = (data) => {
     if (dataToEdit?.id) {
       handleUpdateExistingProduct(data.id, data)
-      handleCloseModal()
+      setOpenModal(false)
       handleDataToEdit(null)
       return
     }
     handleAddNewProduct(data)
-    handleCloseModal()
+    setOpenModal(false)
     handleDataToEdit(null)
   }
 
@@ -71,132 +70,134 @@ const ProductFormModal = ({
             <Error error={modalErrorMsg} />
           </div>
         )}
-        <form onSubmit={handleSubmit(handleData)}>
-          <div className='form-sections-container'>
-            <div className='left-section-container'>
-              <h2>Information</h2>
-              <FormInput
-                name='name'
-                type='text'
-                register={register}
-                validationSchema={{ required: 'Name required' }}
-                errors={errors}
-              />
-              <FormInput
-                name='sku'
-                type='text'
-                register={register}
-                validationSchema={{ required: 'SKU required' }}
-                errors={errors}
-              />
-              <FormInput
-                name='description'
-                register={register}
-                validationSchema={{ required: 'Description required' }}
-                errors={errors}
-              />
-              <div className='price-quantity-container'>
-                <div>
-                  <FormInput
-                    name='price'
-                    type='number'
-                    step='any'
-                    register={register}
-                    errors={errors}
-                    className='price-quantity-input'
-                    validationSchema={{
-                      required: 'Price required',
-                      valueAsNumber: true,
-                    }}
-                  />
+        {!modalErrorMsg && (
+          <form onSubmit={handleSubmit(handleData)}>
+            <div className='form-sections-container'>
+              <div className='left-section-container'>
+                <h2>Information</h2>
+                <FormInput
+                  name='name'
+                  type='text'
+                  register={register}
+                  validationSchema={{ required: 'Name required' }}
+                  errors={errors}
+                />
+                <FormInput
+                  name='sku'
+                  type='text'
+                  register={register}
+                  validationSchema={{ required: 'SKU required' }}
+                  errors={errors}
+                />
+                <FormInput
+                  name='description'
+                  register={register}
+                  validationSchema={{ required: 'Description required' }}
+                  errors={errors}
+                />
+                <div className='price-quantity-container'>
+                  <div>
+                    <FormInput
+                      name='price'
+                      type='number'
+                      step='any'
+                      register={register}
+                      errors={errors}
+                      className='price-quantity-input'
+                      validationSchema={{
+                        required: 'Price required',
+                        valueAsNumber: true,
+                      }}
+                    />
+                  </div>
+                  <div>
+                    <FormInput
+                      name='quantity'
+                      type='number'
+                      register={register}
+                      errors={errors}
+                      className='price-quantity-input'
+                      validationSchema={{
+                        required: 'Quantity required',
+                        valueAsNumber: true,
+                      }}
+                    />
+                  </div>
                 </div>
-                <div>
-                  <FormInput
-                    name='quantity'
-                    type='number'
-                    register={register}
-                    errors={errors}
-                    className='price-quantity-input'
-                    validationSchema={{
-                      required: 'Quantity required',
-                      valueAsNumber: true,
-                    }}
-                  />
-                </div>
+                <label
+                  htmlFor='status'
+                  className={errors.status?.message && 'label-error-style'}
+                >
+                  Status
+                </label>
+                <Controller
+                  name='status'
+                  control={control}
+                  rules={registerOptions.status}
+                  render={({ field }) => (
+                    <Select
+                      {...field}
+                      options={OPTIONS}
+                      placeholder='Choose an option'
+                      styles={customStyles}
+                      formatOptionLabel={optionLabel}
+                      id='status'
+                    />
+                  )}
+                />
+                <label
+                  htmlFor='status'
+                  className={
+                    errors.status?.message
+                      ? 'form-error-message'
+                      : 'form-error-message hidden-style'
+                  }
+                >
+                  {errors.status?.message}
+                </label>
               </div>
-              <label
-                htmlFor='status'
-                className={errors.status?.message && 'label-error-style'}
-              >
-                Status
-              </label>
-              <Controller
-                name='status'
-                control={control}
-                rules={registerOptions.status}
-                render={({ field }) => (
-                  <Select
-                    {...field}
-                    options={OPTIONS}
-                    placeholder='Choose an option'
-                    styles={customStyles}
-                    formatOptionLabel={optionLabel}
-                    id='status'
-                  />
-                )}
-              />
-              <label
-                htmlFor='status'
-                className={
-                  errors.status?.message
-                    ? 'form-error-message'
-                    : 'form-error-message hidden-style'
-                }
-              >
-                {errors.status?.message}
-              </label>
+              <div className='right-section-container'>
+                <h2>Photo</h2>
+                <FormInput
+                  name='imgUrl'
+                  type='text'
+                  register={register}
+                  errors={errors}
+                  validationSchema={{
+                    required: 'Product image URL required',
+                    pattern: {
+                      value: /^(ftp|http|https):\/\/[^ "]+$/,
+                      message: 'Enter a valid image url',
+                    },
+                  }}
+                />
+                <section className='image-preview-container'>
+                  <p className='preview-heading'>Preview</p>
+                  <div className='image-preview-container'>
+                    <img
+                      src={imageProps[0] ? imageProps[0] : placeHolderImg}
+                      alt={imageProps[0] ? imageProps[1] : 'placeholder'}
+                      className='image-preview'
+                      onError={handleImgError}
+                    />
+                  </div>
+                </section>
+              </div>
             </div>
-            <div className='right-section-container'>
-              <h2>Photo</h2>
-              <FormInput
-                name='imgUrl'
-                type='text'
-                register={register}
-                errors={errors}
-                validationSchema={{
-                  required: 'Product image URL required',
-                  pattern: {
-                    value: /^(ftp|http|https):\/\/[^ "]+$/,
-                    message: 'Enter a valid image url',
-                  },
-                }}
+            <div className='buttons-container-form'>
+              <Button
+                onClick={handleFormEditValues}
+                name='Cancel'
+                className='form-button product-form-button-light'
               />
-              <section className='image-preview-container'>
-                <p className='preview-heading'>Preview</p>
-                <div className='image-preview-container'>
-                  <img
-                    src={imageProps[0] ? imageProps[0] : placeHolderImg}
-                    alt={imageProps[0] ? imageProps[1] : 'placeholder'}
-                    className='image-preview'
-                    onError={handleImgError}
-                  />
-                </div>
-              </section>
+              <Button
+                type='submit'
+                name='Save'
+                className='form-button product-form-button-dark'
+              />
             </div>
-          </div>
-          <div className='buttons-container-form'>
-            <Button
-              onClick={handleFormEditValues}
-              name='Cancel'
-              className='form-button product-form-button-light'
-            />
-            <Button
-              type='submit'
-              name='Save'
-              className='form-button product-form-button-dark'
-            />
-          </div>
-        </form>
+          </form>
+        )}
       </section>
     </>
   )
